Move luacfg closer to statreceiver
