+++ cmd/statreceiver/luacfg/cfg_test.go
@@ -1,20 +1,23 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package main
+package luacfg_test
 
 import (
+	"bytes"
 	"fmt"
-	"os"
 
-	"storj.io/storj/pkg/luacfg"
+	"storj.io/storj/cmd/statreceiver/luacfg"
 )
 
-func main() {
+func Example() {
 	s := luacfg.NewScope()
 	s.RegisterVal("print", fmt.Println)
-	err := s.Run(os.Stdin)
+
+	err := s.Run(bytes.NewBufferString(`print "hello"`))
 	if err != nil {
 		panic(err)
 	}
+
+	// Output: hello
 }
